fix test lints
