+++ converter_test.go
@@ -5,6 +5,7 @@ package gosnowflake
 import (
 	"database/sql/driver"
 	"fmt"
+	"github.com/apache/arrow/go/arrow/decimal128"
 	"math/big"
 	"math/cmplx"
 	"reflect"
@@ -15,6 +16,32 @@ import (
 	"github.com/apache/arrow/go/arrow/array"
 	"github.com/apache/arrow/go/arrow/memory"
 )
+
+func stringIntToDecimal(src string) (decimal128.Num, bool) {
+	b, ok := new(big.Int).SetString(src, 10)
+	if !ok {
+		return decimal128.Num{}, ok
+	}
+	var high, low big.Int
+	high.QuoRem(b, decimalShift, &low)
+	return decimal128.New(high.Int64(), low.Uint64()), ok
+}
+
+func stringFloatToDecimal(src string, scale int64) (decimal128.Num, bool) {
+	b, ok := new(big.Float).SetString(src)
+	if !ok {
+		return decimal128.Num{}, ok
+	}
+	s := new(big.Float).SetInt(new(big.Int).Exp(big.NewInt(10), big.NewInt(scale), nil))
+	n := new(big.Float).Mul(b, s)
+	if !n.IsInt() {
+		return decimal128.Num{}, false
+	}
+	var high, low, z big.Int
+	n.Int(&z)
+	high.QuoRem(&z, decimalShift, &low)
+	return decimal128.New(high.Int64(), low.Uint64()), ok
+}
 
 type tcGoTypeToSnowflake struct {
 	in    interface{}
